[SPARK-44343][CONNECT] Prepare ScalaReflection to the move to SQL/API
### What changes were proposed in this pull request?
This PR moves all catalyst specific internals out of ScalaReflection into other catalyst classes:
- Serializer Expression Generation is moved to `SerializerBuildHelper`.
- Deaerializer Expression Generation is moved to `DeserializerBuildHelper`.
- Common utils are moved to `EncoderUtils`.

### Why are the changes needed?
We want to use ScalaReflection based encoder inference both for SQL/Core and Connect.

### Does this PR introduce _any_ user-facing change?
No.

### How was this patch tested?
Existing tests.

Closes #41920 from hvanhovell/SPARK-44343.

Authored-by: Herman van Hovell <herman@databricks.com>
