+++ sql/catalyst/src/main/scala/org/apache/spark/sql/catalyst/ScalaReflection.scala
@@ -20,7 +20,6 @@ package org.apache.spark.sql.catalyst
 import javax.lang.model.SourceVersion
 
 import scala.annotation.tailrec
-import scala.language.existentials
 import scala.reflect.ClassTag
 import scala.reflect.internal.Symbols
 import scala.util.{Failure, Success}
@@ -30,23 +29,13 @@ import org.apache.commons.lang3.reflect.ConstructorUtils
 import org.apache.spark.SPARK_DOC_ROOT
 import org.apache.spark.internal.Logging
 import org.apache.spark.sql.Row
-import org.apache.spark.sql.catalyst.{expressions => exprs}
-import org.apache.spark.sql.catalyst.DeserializerBuildHelper._
-import org.apache.spark.sql.catalyst.SerializerBuildHelper._
-import org.apache.spark.sql.catalyst.analysis.GetColumnByOrdinal
 import org.apache.spark.sql.catalyst.encoders.AgnosticEncoder
 import org.apache.spark.sql.catalyst.encoders.AgnosticEncoders._
-import org.apache.spark.sql.catalyst.expressions._
-import org.apache.spark.sql.catalyst.expressions.objects._
-import org.apache.spark.sql.catalyst.types.{PhysicalBinaryType, PhysicalIntegerType, PhysicalLongType}
-import org.apache.spark.sql.catalyst.util.{ArrayBasedMapData, ArrayData, MapData}
 import org.apache.spark.sql.errors.QueryExecutionErrors
 import org.apache.spark.sql.types._
-import org.apache.spark.unsafe.types.{CalendarInterval, UTF8String}
-
+import org.apache.spark.unsafe.types.CalendarInterval
 
 private[catalyst] object ScalaSubtypeLock
-
 
 /**
  * A default version of ScalaReflection that uses the runtime universe.
@@ -79,489 +68,11 @@ object ScalaReflection extends ScalaReflection {
     }
   }
 
-  /**
-   * Return the data type we expect to see when deserializing a value with encoder `enc`.
-   */
-  private[catalyst] def externalDataTypeFor(enc: AgnosticEncoder[_]): DataType = {
-    externalDataTypeFor(enc, lenientSerialization = false)
-  }
-
-  private[catalyst]  def lenientExternalDataTypeFor(enc: AgnosticEncoder[_]): DataType =
-    externalDataTypeFor(enc, enc.lenientSerialization)
-
-  private def externalDataTypeFor(
-      enc: AgnosticEncoder[_],
-      lenientSerialization: Boolean): DataType = {
-    // DataType can be native.
-    if (isNativeEncoder(enc)) {
-      enc.dataType
-    } else if (lenientSerialization) {
-      ObjectType(classOf[java.lang.Object])
-    } else {
-      ObjectType(enc.clsTag.runtimeClass)
-    }
-  }
-
-  /**
-   * Returns true if the value of this data type is same between internal and external.
-   */
-  def isNativeType(dt: DataType): Boolean = dt match {
-    case NullType | BooleanType | ByteType | ShortType | IntegerType | LongType |
-         FloatType | DoubleType | BinaryType | CalendarIntervalType => true
-    case _ => false
-  }
-
   private def baseType(tpe: `Type`): `Type` = {
     tpe.dealias match {
       case annotatedType: AnnotatedType => annotatedType.underlying
       case other => other
     }
-  }
-
-  /**
-   * Returns true if the encoders' internal and external data type is the same.
-   */
-  private def isNativeEncoder(enc: AgnosticEncoder[_]): Boolean = enc match {
-    case PrimitiveBooleanEncoder => true
-    case PrimitiveByteEncoder => true
-    case PrimitiveShortEncoder => true
-    case PrimitiveIntEncoder => true
-    case PrimitiveLongEncoder => true
-    case PrimitiveFloatEncoder => true
-    case PrimitiveDoubleEncoder => true
-    case NullEncoder => true
-    case CalendarIntervalEncoder => true
-    case BinaryEncoder => true
-    case _: SparkDecimalEncoder => true
-    case _ => false
-  }
-
-  private def toArrayMethodName(enc: AgnosticEncoder[_]): String = enc match {
-    case PrimitiveBooleanEncoder => "toBooleanArray"
-    case PrimitiveByteEncoder => "toByteArray"
-    case PrimitiveShortEncoder => "toShortArray"
-    case PrimitiveIntEncoder => "toIntArray"
-    case PrimitiveLongEncoder => "toLongArray"
-    case PrimitiveFloatEncoder => "toFloatArray"
-    case PrimitiveDoubleEncoder => "toDoubleArray"
-    case _ => "array"
-  }
-
-  /**
-   * Returns an expression for deserializing the Spark SQL representation of an object into its
-   * external form. The mapping between the internal and external representations is
-   * described by encoder `enc`. The Spark SQL representation is located at ordinal 0 of
-   * a row, i.e., `GetColumnByOrdinal(0, _)`. Nested classes will have their fields accessed using
-   * `UnresolvedExtractValue`.
-   *
-   * The returned expression is used by `ExpressionEncoder`. The encoder will resolve and bind this
-   * deserializer expression when using it.
-   *
-   * @param enc encoder that describes the mapping between the Spark SQL representation and the
-   *            external representation.
-   */
-  def deserializerFor[T](enc: AgnosticEncoder[T]): Expression = {
-    val walkedTypePath = WalkedTypePath().recordRoot(enc.clsTag.runtimeClass.getName)
-    // Assumes we are deserializing the first column of a row.
-    val input = GetColumnByOrdinal(0, enc.dataType)
-    enc match {
-      case RowEncoder(fields) =>
-        val children = fields.zipWithIndex.map { case (f, i) =>
-          deserializerFor(f.enc, GetStructField(input, i), walkedTypePath)
-        }
-        CreateExternalRow(children, enc.schema)
-      case _ =>
-        val deserializer = deserializerFor(
-          enc,
-          upCastToExpectedType(input, enc.dataType, walkedTypePath),
-          walkedTypePath)
-        expressionWithNullSafety(deserializer, enc.nullable, walkedTypePath)
-    }
-  }
-
-  /**
-   * Returns an expression for deserializing the value of an input expression into its external
-   * representation. The mapping between the internal and external representations is
-   * described by encoder `enc`.
-   *
-   * @param enc encoder that describes the mapping between the Spark SQL representation and the
-   *            external representation.
-   * @param path The expression which can be used to extract serialized value.
-   * @param walkedTypePath The paths from top to bottom to access current field when deserializing.
-   */
-  private def deserializerFor(
-      enc: AgnosticEncoder[_],
-      path: Expression,
-      walkedTypePath: WalkedTypePath): Expression = enc match {
-    case _ if isNativeEncoder(enc) =>
-      path
-    case _: BoxedLeafEncoder[_, _] =>
-      createDeserializerForTypesSupportValueOf(path, enc.clsTag.runtimeClass)
-    case JavaEnumEncoder(tag) =>
-      val toString = createDeserializerForString(path, returnNullable = false)
-      createDeserializerForTypesSupportValueOf(toString, tag.runtimeClass)
-    case ScalaEnumEncoder(parent, tag) =>
-      StaticInvoke(
-        parent,
-        ObjectType(tag.runtimeClass),
-        "withName",
-        createDeserializerForString(path, returnNullable = false) :: Nil,
-        returnNullable = false)
-    case StringEncoder =>
-      createDeserializerForString(path, returnNullable = false)
-    case _: ScalaDecimalEncoder =>
-      createDeserializerForScalaBigDecimal(path, returnNullable = false)
-    case _: JavaDecimalEncoder =>
-      createDeserializerForJavaBigDecimal(path, returnNullable = false)
-    case ScalaBigIntEncoder =>
-      createDeserializerForScalaBigInt(path)
-    case JavaBigIntEncoder =>
-      createDeserializerForJavaBigInteger(path, returnNullable = false)
-    case DayTimeIntervalEncoder =>
-      createDeserializerForDuration(path)
-    case YearMonthIntervalEncoder =>
-      createDeserializerForPeriod(path)
-    case _: DateEncoder =>
-      createDeserializerForSqlDate(path)
-    case _: LocalDateEncoder =>
-      createDeserializerForLocalDate(path)
-    case _: TimestampEncoder =>
-      createDeserializerForSqlTimestamp(path)
-    case _: InstantEncoder =>
-      createDeserializerForInstant(path)
-    case LocalDateTimeEncoder =>
-      createDeserializerForLocalDateTime(path)
-    case UDTEncoder(udt, udtClass) =>
-      val obj = NewInstance(udtClass, Nil, ObjectType(udtClass))
-      Invoke(obj, "deserialize", ObjectType(udt.userClass), path :: Nil)
-    case OptionEncoder(valueEnc) =>
-      val newTypePath = walkedTypePath.recordOption(valueEnc.clsTag.runtimeClass.getName)
-      val deserializer = deserializerFor(valueEnc, path, newTypePath)
-      WrapOption(deserializer, externalDataTypeFor(valueEnc))
-
-    case ArrayEncoder(elementEnc: AgnosticEncoder[_], containsNull) =>
-      Invoke(
-        deserializeArray(
-          path,
-          elementEnc,
-          containsNull,
-          None,
-          walkedTypePath),
-        toArrayMethodName(elementEnc),
-        ObjectType(enc.clsTag.runtimeClass),
-        returnNullable = false)
-
-    case IterableEncoder(clsTag, elementEnc, containsNull, _) =>
-      deserializeArray(
-        path,
-        elementEnc,
-        containsNull,
-        Option(clsTag.runtimeClass),
-        walkedTypePath)
-
-    case MapEncoder(tag, keyEncoder, valueEncoder, _)
-        if classOf[java.util.Map[_, _]].isAssignableFrom(tag.runtimeClass) =>
-      // TODO (hvanhovell) this is can be improved.
-      val newTypePath = walkedTypePath.recordMap(
-        keyEncoder.clsTag.runtimeClass.getName,
-        valueEncoder.clsTag.runtimeClass.getName)
-
-      val keyData =
-        Invoke(
-          UnresolvedMapObjects(
-            p => deserializerFor(keyEncoder, p, newTypePath),
-            MapKeys(path)),
-          "array",
-          ObjectType(classOf[Array[Any]]))
-
-      val valueData =
-        Invoke(
-          UnresolvedMapObjects(
-            p => deserializerFor(valueEncoder, p, newTypePath),
-            MapValues(path)),
-          "array",
-          ObjectType(classOf[Array[Any]]))
-
-      StaticInvoke(
-        ArrayBasedMapData.getClass,
-        ObjectType(classOf[java.util.Map[_, _]]),
-        "toJavaMap",
-        keyData :: valueData :: Nil,
-        returnNullable = false)
-
-    case MapEncoder(tag, keyEncoder, valueEncoder, _) =>
-      val newTypePath = walkedTypePath.recordMap(
-        keyEncoder.clsTag.runtimeClass.getName,
-        valueEncoder.clsTag.runtimeClass.getName)
-      UnresolvedCatalystToExternalMap(
-        path,
-        deserializerFor(keyEncoder, _, newTypePath),
-        deserializerFor(valueEncoder, _, newTypePath),
-        tag.runtimeClass)
-
-    case ProductEncoder(tag, fields) =>
-      val cls = tag.runtimeClass
-      val dt = ObjectType(cls)
-      val isTuple = cls.getName.startsWith("scala.Tuple")
-      val arguments = fields.zipWithIndex.map {
-        case (field, i) =>
-          val newTypePath = walkedTypePath.recordField(
-            field.enc.clsTag.runtimeClass.getName,
-            field.name)
-          // For tuples, we grab the inner fields by ordinal instead of name.
-          val getter = if (isTuple) {
-            addToPathOrdinal(path, i, field.enc.dataType, newTypePath)
-          } else {
-            addToPath(path, field.name, field.enc.dataType, newTypePath)
-          }
-          expressionWithNullSafety(
-            deserializerFor(field.enc, getter, newTypePath),
-            field.enc.nullable,
-            newTypePath)
-      }
-      expressions.If(
-        IsNull(path),
-        expressions.Literal.create(null, dt),
-        NewInstance(cls, arguments, dt, propagateNull = false))
-
-    case RowEncoder(fields) =>
-      val convertedFields = fields.zipWithIndex.map { case (f, i) =>
-        val newTypePath = walkedTypePath.recordField(
-          f.enc.clsTag.runtimeClass.getName,
-          f.name)
-        exprs.If(
-          Invoke(path, "isNullAt", BooleanType, exprs.Literal(i) :: Nil),
-          exprs.Literal.create(null, externalDataTypeFor(f.enc)),
-          deserializerFor(f.enc, GetStructField(path, i), newTypePath))
-      }
-      exprs.If(IsNull(path),
-        exprs.Literal.create(null, externalDataTypeFor(enc)),
-        CreateExternalRow(convertedFields, enc.schema))
-
-    case JavaBeanEncoder(tag, fields) =>
-      val setters = fields.map { f =>
-        val newTypePath = walkedTypePath.recordField(
-          f.enc.clsTag.runtimeClass.getName,
-          f.name)
-        val setter = expressionWithNullSafety(
-          deserializerFor(
-            f.enc,
-            addToPath(path, f.name, f.enc.dataType, newTypePath),
-            newTypePath),
-          nullable = f.nullable,
-          newTypePath)
-        f.writeMethod.get -> setter
-      }
-
-      val cls = tag.runtimeClass
-      val newInstance = NewInstance(cls, Nil, ObjectType(cls), propagateNull = false)
-      val result = InitializeJavaBean(newInstance, setters.toMap)
-      exprs.If(IsNull(path), exprs.Literal.create(null, ObjectType(cls)), result)
-  }
-
-  private def deserializeArray(
-      path: Expression,
-      elementEnc: AgnosticEncoder[_],
-      containsNull: Boolean,
-      cls: Option[Class[_]],
-      walkedTypePath: WalkedTypePath): Expression = {
-    val newTypePath = walkedTypePath.recordArray(elementEnc.clsTag.runtimeClass.getName)
-    val mapFunction: Expression => Expression = element => {
-      // upcast the array element to the data type the encoder expects.
-      deserializerForWithNullSafetyAndUpcast(
-        element,
-        elementEnc.dataType,
-        nullable = containsNull,
-        newTypePath,
-        deserializerFor(elementEnc, _, newTypePath))
-    }
-    UnresolvedMapObjects(mapFunction, path, cls)
-  }
-
-  /**
-   * Returns an expression for serializing an object into its Spark SQL form. The mapping
-   * between the external and internal representations is described by encoder `enc`. The
-   * input object is located at ordinal 0 of a row, i.e., `BoundReference(0, _)`.
-   */
-  def serializerFor(enc: AgnosticEncoder[_]): Expression = {
-    val input = BoundReference(0, lenientExternalDataTypeFor(enc), nullable = enc.nullable)
-    serializerFor(enc, input)
-  }
-
-  /**
-   * Returns an expression for serializing the value of an input expression into its Spark SQL
-   * representation. The mapping between the external and internal representations is described
-   * by encoder `enc`.
-   */
-  private def serializerFor(enc: AgnosticEncoder[_], input: Expression): Expression = enc match {
-    case _ if isNativeEncoder(enc) => input
-    case BoxedBooleanEncoder => createSerializerForBoolean(input)
-    case BoxedByteEncoder => createSerializerForByte(input)
-    case BoxedShortEncoder => createSerializerForShort(input)
-    case BoxedIntEncoder => createSerializerForInteger(input)
-    case BoxedLongEncoder => createSerializerForLong(input)
-    case BoxedFloatEncoder => createSerializerForFloat(input)
-    case BoxedDoubleEncoder => createSerializerForDouble(input)
-    case JavaEnumEncoder(_) => createSerializerForJavaEnum(input)
-    case ScalaEnumEncoder(_, _) => createSerializerForScalaEnum(input)
-    case StringEncoder => createSerializerForString(input)
-    case ScalaDecimalEncoder(dt) => createSerializerForBigDecimal(input, dt)
-    case JavaDecimalEncoder(dt, false) => createSerializerForBigDecimal(input, dt)
-    case JavaDecimalEncoder(dt, true) => createSerializerForAnyDecimal(input, dt)
-    case ScalaBigIntEncoder => createSerializerForBigInteger(input)
-    case JavaBigIntEncoder => createSerializerForBigInteger(input)
-    case DayTimeIntervalEncoder => createSerializerForJavaDuration(input)
-    case YearMonthIntervalEncoder => createSerializerForJavaPeriod(input)
-    case DateEncoder(true) | LocalDateEncoder(true) => createSerializerForAnyDate(input)
-    case DateEncoder(false) => createSerializerForSqlDate(input)
-    case LocalDateEncoder(false) => createSerializerForJavaLocalDate(input)
-    case TimestampEncoder(true) | InstantEncoder(true) => createSerializerForAnyTimestamp(input)
-    case TimestampEncoder(false) => createSerializerForSqlTimestamp(input)
-    case InstantEncoder(false) => createSerializerForJavaInstant(input)
-    case LocalDateTimeEncoder => createSerializerForLocalDateTime(input)
-    case UDTEncoder(udt, udtClass) => createSerializerForUserDefinedType(input, udt, udtClass)
-    case OptionEncoder(valueEnc) =>
-      serializerFor(valueEnc, UnwrapOption(externalDataTypeFor(valueEnc), input))
-
-    case ArrayEncoder(elementEncoder, containsNull) =>
-      if (elementEncoder.isPrimitive) {
-        createSerializerForPrimitiveArray(input, elementEncoder.dataType)
-      } else {
-        serializerForArray(elementEncoder, containsNull, input, lenientSerialization = false)
-      }
-
-    case IterableEncoder(ctag, elementEncoder, containsNull, lenientSerialization) =>
-      val getter = if (classOf[scala.collection.Set[_]].isAssignableFrom(ctag.runtimeClass)) {
-        // There's no corresponding Catalyst type for `Set`, we serialize a `Set` to Catalyst array.
-        // Note that the property of `Set` is only kept when manipulating the data as domain object.
-        Invoke(input, "toSeq", ObjectType(classOf[scala.collection.Seq[_]]))
-      } else {
-        input
-      }
-      serializerForArray(elementEncoder, containsNull, getter, lenientSerialization)
-
-    case MapEncoder(_, keyEncoder, valueEncoder, valueContainsNull) =>
-      createSerializerForMap(
-        input,
-        MapElementInformation(
-          ObjectType(classOf[AnyRef]),
-          nullable = keyEncoder.nullable,
-          validateAndSerializeElement(keyEncoder, keyEncoder.nullable)),
-        MapElementInformation(
-          ObjectType(classOf[AnyRef]),
-          nullable = valueContainsNull,
-          validateAndSerializeElement(valueEncoder, valueContainsNull))
-      )
-
-    case ProductEncoder(_, fields) =>
-      val serializedFields = fields.map { field =>
-        // SPARK-26730 inputObject won't be null with If's guard below. And KnownNotNul
-        // is necessary here. Because for a nullable nested inputObject with struct data
-        // type, e.g. StructType(IntegerType, StringType), it will return nullable=true
-        // for IntegerType without KnownNotNull. And that's what we do not expect to.
-        val getter = Invoke(
-          KnownNotNull(input),
-          field.name,
-          externalDataTypeFor(field.enc),
-          returnNullable = field.nullable)
-        field.name -> serializerFor(field.enc, getter)
-      }
-      createSerializerForObject(input, serializedFields)
-
-    case RowEncoder(fields) =>
-      val serializedFields = fields.zipWithIndex.map { case (field, index) =>
-        val fieldValue = serializerFor(
-          field.enc,
-          ValidateExternalType(
-            GetExternalRowField(input, index, field.name),
-            field.enc.dataType,
-            lenientExternalDataTypeFor(field.enc)))
-
-        val convertedField = if (field.nullable) {
-          exprs.If(
-            Invoke(input, "isNullAt", BooleanType, exprs.Literal(index) :: Nil),
-            // Because we strip UDTs, `field.dataType` can be different from `fieldValue.dataType`.
-            // We should use `fieldValue.dataType` here.
-            exprs.Literal.create(null, fieldValue.dataType),
-            fieldValue
-          )
-        } else {
-          AssertNotNull(fieldValue)
-        }
-        field.name -> convertedField
-      }
-      createSerializerForObject(input, serializedFields)
-
-    case JavaBeanEncoder(_, fields) =>
-      val serializedFields = fields.map { f =>
-        val fieldValue = Invoke(
-          KnownNotNull(input),
-          f.readMethod.get,
-          externalDataTypeFor(f.enc),
-          propagateNull = f.nullable,
-          returnNullable = f.nullable)
-        f.name -> serializerFor(f.enc, fieldValue)
-      }
-      createSerializerForObject(input, serializedFields)
-  }
-
-  private def serializerForArray(
-      elementEnc: AgnosticEncoder[_],
-      elementNullable: Boolean,
-      input: Expression,
-      lenientSerialization: Boolean): Expression = {
-    // Default serializer for Seq and generic Arrays. This does not work for primitive arrays.
-    val genericSerializer = createSerializerForMapObjects(
-      input,
-      ObjectType(classOf[AnyRef]),
-      validateAndSerializeElement(elementEnc, elementNullable))
-
-    // Check if it is possible the user can pass a primitive array. This is the only case when it
-    // is safe to directly convert to an array (for generic arrays and Seqs the type and the
-    // nullability can be violated). If the user has passed a primitive array we create a special
-    // code path to deal with these.
-    val primitiveEncoderOption = elementEnc match {
-      case _ if !lenientSerialization => None
-      case enc: PrimitiveLeafEncoder[_] => Option(enc)
-      case enc: BoxedLeafEncoder[_, _] => Option(enc.primitive)
-      case _ => None
-    }
-    primitiveEncoderOption match {
-      case Some(primitiveEncoder) =>
-        val primitiveArrayClass = primitiveEncoder.clsTag.wrap.runtimeClass
-        val check = Invoke(
-          targetObject = exprs.Literal.fromObject(primitiveArrayClass),
-          functionName = "isInstance",
-          BooleanType,
-          arguments = input :: Nil,
-          propagateNull = false,
-          returnNullable = false)
-        exprs.If(
-          check,
-          // TODO replace this with `createSerializerForPrimitiveArray` as
-          //  soon as Cast support ObjectType casts.
-          StaticInvoke(
-            classOf[ArrayData],
-            ArrayType(elementEnc.dataType, containsNull = false),
-            "toArrayData",
-            input :: Nil,
-            propagateNull = false,
-            returnNullable = false),
-          genericSerializer)
-      case None =>
-        genericSerializer
-    }
-  }
-
-  private def validateAndSerializeElement(
-      enc: AgnosticEncoder[_],
-      nullable: Boolean): Expression => Expression = { input =>
-    expressionWithNullSafety(
-      serializerFor(
-        enc,
-        ValidateExternalType(input, enc.dataType, lenientExternalDataTypeFor(enc))),
-      nullable,
-      WalkedTypePath())
   }
 
   /**
@@ -606,15 +117,6 @@ object ScalaReflection extends ScalaReflection {
     }
   }
 
-  /**
-   * Returns the parameter values for the primary constructor of this class.
-   */
-  def getConstructorParameterValues(obj: DefinedByConstructorParams): Seq[AnyRef] = {
-    getConstructorParameterNames(obj.getClass).map { name =>
-      obj.getClass.getMethod(name).invoke(obj)
-    }
-  }
-
   private def erasure(tpe: Type): Type = {
     // For user-defined AnyVal classes, we should not erasure it. Otherwise, it will
     // resolve to underlying type which wrapped by this class, e.g erasure
@@ -650,13 +152,6 @@ object ScalaReflection extends ScalaReflection {
     mirror.runtimeClass(erasure(tpe).dealias.typeSymbol.asClass)
 
   case class Schema(dataType: DataType, nullable: Boolean)
-
-  /** Returns a Sequence of attributes for the given case class type. */
-  def attributesFor[T: TypeTag]: Seq[Attribute] = schemaFor[T] match {
-    case Schema(s: StructType, _) =>
-      s.toAttributes
-    case others => throw QueryExecutionErrors.attributesForTypeUnsupportedError(others)
-  }
 
   /** Returns a catalyst DataType and its nullability for the given Scala Type using reflection. */
   def schemaFor[T: TypeTag]: Schema = schemaFor(localTypeOf[T])
@@ -710,72 +205,6 @@ object ScalaReflection extends ScalaReflection {
       case t if isSubtype(t, localTypeOf[Option[_]]) => definedByConstructorParams(t.typeArgs.head)
       case _ => isSubtype(tpe.dealias, localTypeOf[Product]) ||
         isSubtype(tpe.dealias, localTypeOf[DefinedByConstructorParams])
-    }
-  }
-
-  val typeJavaMapping: Map[DataType, Class[_]] = Map[DataType, Class[_]](
-    BooleanType -> classOf[Boolean],
-    ByteType -> classOf[Byte],
-    ShortType -> classOf[Short],
-    IntegerType -> classOf[Int],
-    LongType -> classOf[Long],
-    FloatType -> classOf[Float],
-    DoubleType -> classOf[Double],
-    StringType -> classOf[UTF8String],
-    DateType -> classOf[PhysicalIntegerType.InternalType],
-    TimestampType -> classOf[PhysicalLongType.InternalType],
-    TimestampNTZType -> classOf[PhysicalLongType.InternalType],
-    BinaryType -> classOf[PhysicalBinaryType.InternalType],
-    CalendarIntervalType -> classOf[CalendarInterval]
-  )
-
-  val typeBoxedJavaMapping: Map[DataType, Class[_]] = Map[DataType, Class[_]](
-    BooleanType -> classOf[java.lang.Boolean],
-    ByteType -> classOf[java.lang.Byte],
-    ShortType -> classOf[java.lang.Short],
-    IntegerType -> classOf[java.lang.Integer],
-    LongType -> classOf[java.lang.Long],
-    FloatType -> classOf[java.lang.Float],
-    DoubleType -> classOf[java.lang.Double],
-    DateType -> classOf[java.lang.Integer],
-    TimestampType -> classOf[java.lang.Long],
-    TimestampNTZType -> classOf[java.lang.Long]
-  )
-
-  def dataTypeJavaClass(dt: DataType): Class[_] = {
-    dt match {
-      case _: DecimalType => classOf[Decimal]
-      case _: DayTimeIntervalType => classOf[PhysicalLongType.InternalType]
-      case _: YearMonthIntervalType => classOf[PhysicalIntegerType.InternalType]
-      case _: StructType => classOf[InternalRow]
-      case _: ArrayType => classOf[ArrayData]
-      case _: MapType => classOf[MapData]
-      case ObjectType(cls) => cls
-      case _ => typeJavaMapping.getOrElse(dt, classOf[java.lang.Object])
-    }
-  }
-
-  @scala.annotation.tailrec
-  def javaBoxedType(dt: DataType): Class[_] = dt match {
-    case _: DecimalType => classOf[Decimal]
-    case _: DayTimeIntervalType => classOf[java.lang.Long]
-    case _: YearMonthIntervalType => classOf[java.lang.Integer]
-    case BinaryType => classOf[Array[Byte]]
-    case StringType => classOf[UTF8String]
-    case CalendarIntervalType => classOf[CalendarInterval]
-    case _: StructType => classOf[InternalRow]
-    case _: ArrayType => classOf[ArrayData]
-    case _: MapType => classOf[MapData]
-    case udt: UserDefinedType[_] => javaBoxedType(udt.sqlType)
-    case ObjectType(cls) => cls
-    case _ => ScalaReflection.typeBoxedJavaMapping.getOrElse(dt, classOf[java.lang.Object])
-  }
-
-  def expressionJavaClasses(arguments: Seq[Expression]): Seq[Class[_]] = {
-    if (arguments != Nil) {
-      arguments.map(e => dataTypeJavaClass(e.dataType))
-    } else {
-      Seq.empty
     }
   }
